Migrate CPU sklearn-like models to per-model folder layout
Apply the TabSTAR Stage 1 recipe to ebm, knn, lr, and perpetual_booster.
Each gets `hpo.py` (search space + `generate_configs_*` helpers) and
`info.py` (MethodMetadata + ModelInfo bundle); `generate.py` becomes a
thin shim so `models/utils.py:name_to_import_map` keeps working.

The `ebm_metadata`, `knn_metadata`, `lr_metadata`, and
`perpetualbooster_metadata` entries in the dated metadata files are now
sourced from each model's `info.py` (legacy names kept as aliases).
