+++ tabarena/tabarena/models/ebm/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from tabarena.models.ebm.hpo import gen_ebm
+from tabarena.models.ebm.info import ebm_info, ebm_method_metadata
+
+__all__ = ["ebm_info", "ebm_method_metadata", "gen_ebm"]
