+++ tabarena/tabarena/models/knn/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from tabarena.models.knn.hpo import gen_knn
+from tabarena.models.knn.info import knn_info, knn_method_metadata
+
+__all__ = ["gen_knn", "knn_info", "knn_method_metadata"]
