+++ tabarena/tabarena/models/lr/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from tabarena.models.lr.hpo import gen_linear
+from tabarena.models.lr.info import lr_info, lr_method_metadata
+
+__all__ = ["gen_linear", "lr_info", "lr_method_metadata"]
