+++ tabarena/tabarena/models/perpetual_booster/__init__.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from tabarena.models.perpetual_booster.hpo import gen_perpetual_booster
+from tabarena.models.perpetual_booster.info import (
+    perpetual_booster_info,
+    perpetual_booster_method_metadata,
+)
+
+__all__ = [
+    "gen_perpetual_booster",
+    "perpetual_booster_info",
+    "perpetual_booster_method_metadata",
+]
