+++ tabarena/tabarena/nips2025_utils/artifacts/_tabarena_method_metadata_2025_09_03.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 
+from tabarena.models.ebm.info import ebm_method_metadata
 from tabarena.models.mitra.info import mitra_method_metadata
 from tabarena.models.realmlp.info import realmlp_method_metadata
 from tabarena.models.xrfm.info import xrfm_method_metadata
 from tabarena.nips2025_utils.artifacts.method_metadata import MethodMetadata
 
 # Legacy aliases — preserved for any external code still importing the older names.
+ebm_metadata = ebm_method_metadata
 mitra_metadata = mitra_method_metadata
 realmlp_gpu_metadata = realmlp_method_metadata
 xrfm_metadata = xrfm_method_metadata
@@ -45,20 +47,6 @@ limix_metadata = MethodMetadata(
     is_bag=False,
     verified=False,
     reference_url="https://arxiv.org/abs/2509.03505",
-    **_common_kwargs,
-)
-ebm_metadata = MethodMetadata(
-    method="ExplainableBM",
-    method_type="config",
-    display_name="EBM",
-    compute="cpu",
-    date="2025-09-03",
-    ag_key="EBM",
-    config_default="ExplainableBM_c1_BAG_L1",
-    can_hpo=True,
-    is_bag=True,
-    verified=True,
-    reference_url="https://www.cs.cornell.edu/~yinlou/papers/lou-kdd13.pdf",
     **_common_kwargs,
 )
 tabflex_metadata = MethodMetadata(
