+++ tabarena/tabarena/nips2025_utils/artifacts/_tabarena_method_metadata_2026_03_18.py
@@ -1,32 +1,10 @@
+from tabarena.models.perpetual_booster.info import perpetual_booster_method_metadata
 from tabarena.models.tabstar.info import tabstar_method_metadata
 from tabarena.nips2025_utils.artifacts.method_metadata import MethodMetadata
 
+# Legacy aliases — preserved for any external code still importing the older names.
+perpetualbooster_metadata = perpetual_booster_method_metadata
 tabstar_metadata = tabstar_method_metadata
-
-
-perpetualbooster_metadata = MethodMetadata(
-    method="PerpetualBooster",
-    method_type="config",
-    display_name="PerpetualBooster",
-    compute="cpu",
-    date="2026-03-06",
-    ag_key="PB",
-    model_key="PB",
-    config_default="PerpetualBooster_c1_BAG_L1",
-    can_hpo=True,
-    is_bag=True,
-    has_raw=True,
-    has_processed=True,
-    artifact_name="tabarena-2026-03-18",
-    s3_bucket="tabarena",
-    s3_prefix="cache",
-    upload_as_public=True,
-    has_results=True,
-    name_suffix=None,
-    verified=True,
-    reference_url="https://perpetual-ml.com/",
-    cache_type="r2",
-)
 
 tabpfn26_metadata = MethodMetadata(
     method="TabPFN-v2.6",
